requirements.txt
altair==4.1.0
bs4==0.0.1
# flake8==3.9.2
fonttools==4.28.4
matplotlib==3.5.1
MultiApp==1.0
numpy==1.20.3
pandas==1.3.5
pep8==1.7.1
Pillow==8.4.0
pyarrow==6.0.1
pylint==2.10.2
seaborn==0.11.2
shap==0.40.0
sklearn==0.0
streamlit==1.3.0
xlrd==2.0.1
yfinance==0.1.67
protobuf==3.20.0 # due to the following error https://stackoverflow.com/questions/72441758/typeerror-descriptors-cannot-not-be-created-directly/72493690#72493690
networkx
ortoolpy
pyvis
